Update chrono.py with optional averaging of E/I datapoints for R_el
Introduced a boolean 'short' (default=True) to the h2x_avg function (and necessarily to the called h2x function). 

If True the linear fit to the chrono data to get the intercept and slope for H2 crossover current and short current(i.e. R_el) are calculated normally. 

If Flase, the datapoints will be averaged to get the H2 crossover current and R_el will be set to np.inf. This is sensible, if the short current is so small that no linear fit can be done. 

In the subsequent data processing be advised to use short=False in the tafel function.